pyproject.toml
[tool.poetry]
name = "calendar"
version = "0.1.0"
description = ""
authors = ["Dao Nguyen <comnuoc@gmail.com>"]
readme = "README.md"
packages = [
    { include = "comnuoc", from = "src" },
]


[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.2"
typing-extensions = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
